refactor: Shifts the responsibility for formula to 'Formula'
