+++ app/src/test/java/com/string/calculator/FormulaTest.java
@@ -1,0 +1,17 @@
+package com.string.calculator;
+
+import org.junit.jupiter.api.Test;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+class FormulaTest {
+  @Test
+  void result() {
+    var formula = new Formula("123 + 23 * 12 - 10 + 1 * 1");
+
+    var actual = formula.result();
+
+    assertThat(actual, is("390"));
+  }
+}
